# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/expenses": {
      "get": {
        "description": "Returns expenses, newest month first, optionally filtered by month range, category, or the import that wrote them.\n\nThis is also the report's drill down. A report cell is one category in one month, which is `categoryId` with `from` and `to` set to the same month, so there is no separate drill down endpoint that could disagree with this one.\n\nFiltering by `importBatchId` answers the question a reviewer actually asks after an import, which is what that file put in.",
        "operationId": "ExpensesController_list_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "How many records to return.",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "How many records to skip.",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "First month of the range, inclusive.",
            "schema": {
              "example": "2026-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Last month of the range, inclusive.",
            "schema": {
              "example": "2026-12",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Restrict to one category.",
            "schema": {
              "example": "65f1c2d3e4b5a6c7d8e9f0a1",
              "type": "string"
            }
          },
          {
            "name": "importBatchId",
            "required": false,
            "in": "query",
            "description": "Restrict to the expenses one import wrote.",
            "schema": {
              "example": "65f1c2d3e4b5a6c7d8e9f0d1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of expenses.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseResponseDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List expenses",
        "tags": [
          "Expenses"
        ]
      },
      "post": {
        "description": "Logs one expense against a category and a month.\n\nEach call creates a record. Logging twice against the same category and month is two expenses, not an overwrite, because that is what spending is: a month's total is the sum of its line items. The report adds them up. This is the opposite of a plan, where one category and month hold a single target.\n\nAn expense belongs to a month rather than a date. Nothing here is a timestamp, so no timezone can move spend into the wrong period.\n\nA negative amount is allowed and means money came back, such as a refund or a credit note.\n\nRejected with `423` when the month is closed, and with `404` when the category is not one the caller can use.",
        "operationId": "ExpensesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The stored expense.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "The category is not one the caller can use.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "423": {
            "description": "The month is closed. The response carries the month in `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log an expense",
        "tags": [
          "Expenses"
        ]
      }
    },
    "/api/v1/expenses/{expenseId}": {
      "patch": {
        "description": "Changes an expense's category, month, amount, or note. Omitted fields are left alone.\n\nThe month is changeable here, unlike on a plan. Spend genuinely lands in the wrong month, usually because an invoice was dated differently from when the cost was incurred, and correcting that is an edit rather than a different record.\n\nThat is why a month change is checked at both ends. Moving an expense out of a closed month changes that month's total, so it is an edit to a closed period just as much as moving one in. A request that touches a closed month at either end is rejected with `423` naming the month that blocked it.",
        "operationId": "ExpensesController_update_v1",
        "parameters": [
          {
            "name": "expenseId",
            "required": true,
            "in": "path",
            "description": "Identifier from the expense list.",
            "schema": {
              "example": "65f1c2d3e4b5a6c7d8e9f0c1",
              "allOf": [
                {
                  "$ref": "#/components/schemas/ObjectId"
                }
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated expense.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "The caller has no expense with that id, or the new category is not one they can use.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "423": {
            "description": "The month is closed. The response carries the month in `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Correct an expense",
        "tags": [
          "Expenses"
        ]
      },
      "delete": {
        "description": "Removes an expense, so it no longer counts toward the actual figure for its category and month.\n\nThe delete is soft. The audit trail keeps what the expense was, so a report run before the deletion can still be explained.\n\nRejected with `423` when the expense's month is closed.",
        "operationId": "ExpensesController_remove_v1",
        "parameters": [
          {
            "name": "expenseId",
            "required": true,
            "in": "path",
            "description": "Identifier from the expense list.",
            "schema": {
              "example": "65f1c2d3e4b5a6c7d8e9f0c1",
              "allOf": [
                {
                  "$ref": "#/components/schemas/ObjectId"
                }
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The expense was removed."
          },
          "401": {
            "description": "The access token is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "The caller has no expense with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "423": {
            "description": "The month is closed. The response carries the month in `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove an expense",
        "tags": [
          "Expenses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Plan vs Actual API",
    "description": "Monthly spending targets, logged expenses, and variance reporting with locked periods.\n\n**Money** is an integer count of minor units. A field ending in `Minor` holds cents.\n\n**Months** are the string `YYYY-MM`.\n\n**Variance** is `actual - plan`. Variance percent is `null` when the plan is zero, never `NaN`.\n\n**Missing actuals** default to `0`. Pass `missingActuals=null` to receive `null` instead. Every report row carries `hasActual`, so a logged zero is never confused with nothing logged.\n\n**Locked periods** reject writes with `423` and the code `PERIOD_LOCKED`.\n\n**Deletes are soft.** A `DELETE` answers `204` and the record stops appearing in reads.\n\n**Sessions** travel in the `Authorization` header, never a cookie, so a mobile or desktop client is a first class caller. Access tokens are RS256 signed and short lived; refresh tokens are opaque, single use, and rotate.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponseDTO": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status, repeated in the body.",
            "example": 423
          },
          "code": {
            "type": "string",
            "description": "Stable code a client branches on. Never parse the message.",
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "IMPORT_VALIDATION_FAILED",
              "PERIOD_LOCKED",
              "RATE_LIMITED",
              "INTERNAL"
            ],
            "example": "PERIOD_LOCKED"
          },
          "message": {
            "type": "string",
            "description": "What went wrong, written for a person.",
            "example": "2026-01 is locked and cannot be edited."
          },
          "details": {
            "type": "object",
            "description": "Machine readable specifics: the offending month, the failing rows, the invalid fields.",
            "example": {
              "month": "2026-01"
            }
          },
          "requestId": {
            "type": "string",
            "description": "Correlates this response with its log lines and the x-request-id header.",
            "example": "0f9c1e2a-7b3d-4c5e-9a8f-1b2c3d4e5f6a"
          },
          "path": {
            "type": "string",
            "description": "The path that produced the error.",
            "example": "/api/v1/plans"
          },
          "timestamp": {
            "type": "string",
            "description": "When the error was produced.",
            "example": "2026-01-15T10:04:11.212Z",
            "format": "date-time"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "requestId",
          "path",
          "timestamp"
        ]
      },
...
      "ObjectId": {
        "type": "object",
        "properties": {}
      },
...
      "ExpenseResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifier of the expense.",
            "example": "65f1c2d3e4b5a6c7d8e9f0c1"
          },
          "categoryId": {
            "type": "string",
            "description": "The category it belongs to.",
            "example": "65f1c2d3e4b5a6c7d8e9f0a1"
          },
          "month": {
            "type": "string",
            "description": "The month it belongs to.",
            "example": "2026-01"
          },
          "amountMinor": {
            "type": "number",
            "description": "The amount in minor units. Negative means money came back.",
            "example": 480000
          },
          "note": {
            "type": "object",
            "description": "Free text.",
            "example": "Annual renewal",
            "nullable": true
          },
          "source": {
            "type": "string",
            "description": "How it got here.",
            "enum": [
              "MANUAL",
              "CSV"
            ],
            "example": "MANUAL"
          },
          "importBatchId": {
            "type": "object",
            "description": "The import that wrote it, when it came from a file.",
            "example": "65f1c2d3e4b5a6c7d8e9f0d1",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "When it was logged.",
            "example": "2026-01-15T10:04:11.212Z",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "categoryId",
          "month",
          "amountMinor",
          "source",
          "createdAt"
        ]
      },
      "CreateExpenseDTO": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "description": "The category the expense belongs to.",
            "example": "65f1c2d3e4b5a6c7d8e9f0a1"
          },
          "month": {
            "type": "string",
            "description": "The month the expense belongs to. An expense belongs to a month, not to a date, so nothing here is a timestamp.",
            "example": "2026-01"
          },
          "amountMinor": {
            "type": "number",
            "description": "The amount in minor units. 4,800.00 is 480000. Negative is allowed and means money came back, such as a refund or a credit note.",
            "example": 480000
          },
          "note": {
            "type": "string",
            "description": "Optional free text.",
            "example": "Annual renewal",
            "maxLength": 280
          }
        },
        "required": [
          "categoryId",
          "month",
          "amountMinor"
        ]
      },
      "UpdateExpenseDTO": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "description": "Move the expense to a different category.",
            "example": "65f1c2d3e4b5a6c7d8e9f0a1"
          },
          "month": {
            "type": "string",
            "description": "Move the expense to a different month. Both the old and the new month must be open.",
            "example": "2026-02"
          },
          "amountMinor": {
            "type": "number",
            "description": "Correct the amount, in minor units.",
            "example": 490000
          },
          "note": {
            "type": "string",
            "description": "Replace the note.",
            "example": "Corrected after credit note",
            "maxLength": 280
          }
        }
      }
    }
  }
}